requirements.txt
Django==5.2.7
djangorestframework==3.16.1
pymongo==4.15.3
dnspython==2.8.0
python-dotenv>=1.0.1
django-cors-headers>=4.5.0
